% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca.coef.R
\name{ca.coef}
\alias{ca.coef}
\title{Coordinate points for a ca object.}
\usage{
ca.coef(x)
}
\arguments{
\item{x}{A ca object.}
}
\value{
The function returns a list with two elements, coordinate points for first
and second dimensions.
}
\description{
This function calculates the coordinate points for an object
with class \code{ca} obtained from a correspondence analysis.
}
