% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_quantiles.R
\name{manual_quantiles}
\alias{manual_quantiles}
\title{Quantiles manually}
\usage{
manual_quantiles(x)
}
\arguments{
\item{x}{numeric vector.}
}
\description{
This function calculates quantiles using the classroom formula.
}
\examples{
# Generating a random vector
x <- rnorm(n=100)
manual_quantiles(x)

}
