% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prof.R
\name{plot_prof}
\alias{plot_prof}
\title{Profile plot}
\usage{
plot_prof(x, Row = TRUE, cex.names = 0.5, xlab = "",
  ylab = "Frequency", cex = 1, ...)
}
\arguments{
\item{x}{A table, xtabs or matrix object.}

\item{Row}{A logical value to indicate if profile plot is for Rows o Columns.}

\item{cex.names}{Expansion factor for axis names (bar labels).}

\item{xlab}{A label for the x axis.}

\item{ylab}{A label for the y axis, by default is Frequency.}

\item{cex}{Character expansion factor for legend.}

\item{...}{Other plotting parameters to affect the plot.}
}
\value{
\code{plot_prof} function returns a profile plot.
}
\description{
This function gives the profile plot for contingency table objects created by table or xtabs or matrix.
}
\examples{
X <- matrix(c(688, 116, 584, 188, 4,
              326, 38, 241, 110, 3,
              343, 84, 909, 412, 26,
              98, 48, 403, 681, 85), nrow=4, byrow=TRUE)

dimnames(X) <- list(eye.color=c("Light","Blue", "Medium","Dark"),
                    hair.color=c("Blond","Red", "Medium","Dark", "Black"))

plot_prof(x=X, Row=TRUE)
plot_prof(x=X, Row=FALSE, cex.names=0.9, cex=0.6)

}
