% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_var_pc.R
\name{plot_var_pc}
\alias{plot_var_pc}
\title{Plot variables in the coordinate plane PC1-PC2.}
\usage{
plot_var_pc(mod, col.arrow = "gray60", xlab = NULL, ylab = NULL, ...)
}
\arguments{
\item{mod}{A prcomp or princomp object.}

\item{col.arrow}{color for the arrows.}

\item{xlab, ylab}{labels for x axis and y axis.}

\item{...}{Arguments to be passed to methods, such as graphical parameters (see par).}
}
\description{
This function plots the variables in the coordinate plane PC1-PC2.
}
\examples{
# For a prcomp object
pca1 <- prcomp(USArrests, scale = TRUE)  # Example
plot_var_pc(pca1)
plot_var_pc(pca1, main='Variables in PC1 and PC2',
           col='blue3', lwd=3)
#
# For a princomp object
pca2 <- princomp(x=USArrests, cor=TRUE)
plot_var_pc(pca2, col='red', lwd=4)

}
\seealso{
[plot_obs_pc()].
}
